Added --with-cooling=CHIMES option.
This option will be used to run with the CHIMES non-equilibrium chemistry and cooling module. In the first instance, I have simply copied over the various header files from src/cooling/none/ into src/cooling/CHIMES/. So currently, this option is identical to --with-cooling=none. The CHIMES functionality will then be built on top of this.