Added core CHIMES files.
The core files for the CHIMES module have been added to the src/cooling/CHIMES/chimes/ directory. I have added these to the src/Makefile.am, and the CHIMES header files have been included in /src/cooling/CHIMES/cooling.h.

I have also updated configure.ac to look for the Sundials libraries, which are required by the CHIMES library. The path to these libraries can be specified via the --with-sundials configure option. If CHIMES has been enabled using --with-cooling=CHIMES but the Sundials library is not present, ./configure will produce an error message.