Update to how CHIMES deals with multiple threads.
The arrays that store the current reaction rates within the CHIMES module have been re-structured to ensure that they are thread-safe. Now, memory is allocated to these arrays every time the chimes_network() routine is called, and then freed at the end of the routine. This ensures that each thread will create its own, local, copy of these arrays.