Added chimes_abundances array for each gas particle.
I have added the chimes_abundances array for each gas particle in the cooling_xpart_data structure, and I have updated the cooling_first_init_part() routine to set the initial values of this array. For now, we will simply set the abundances to be singly ionised. However, in the future we will want to add an option to compute the initial abundances in equilibrium.

I have also added two optional CHIMESCooling parameters, S_over_Si_in_solar and Ca_over_Si_in_solar. These are defined in the same way as for COLIBRE cooling, with a default value of 1.