Added chimes_update_element_abundances() routine.
This routine updates the element abundances in the ChimesGasVars structure based on the particle's mass metal fractions. It then checks whether the total abundance of individual ions/molecules from each element is consistent with the corresponding element abundance. If the latter has changed since the previous call (for example due to enrichment or metal diffusion), then the ion and molecule abundances are adjusted accordingly, such that the relative fractions are preserved. If a metal is now present that wasn't present before (for example, when a primordial particle is first enriched), then that metal is set to fully neutral.