Added new radiation field option.
If UV_field_flag == 1, we use a single, constant radiation field specified through the parameter file. This is controlled by two additional parameters:

PhotonIonTable - name of the HDF5 table containing the photoionisation cross sections data for the given spectrum. This is assumed to be located within the data_path/ directory.
isotropic_photon_density - normalisation of the radiation field, expressed as a photon density (i.e. flux divided by the speed of light), in units of cm^-3. This is kept constant throughout the simulation.