Fixed bug in the CHIMES module.
When self-shielding is disabled, some of the variables within the CHIMES module (e.g. the epsilon's, i.e. the average excess energy per photon) were not being calculated at all, which led to undefined behaviour. We have now fixed this, so that these variables are set to their optically thin values.