Added ChimesAbundances array to I/O.
We have added the ChimesAbundances array to the snapshots through the cooling_write_particles() function in src/cooling/CHIMES/cooling_io.h.

Note that the size of the abundance array is hard-coded to be the maximum size, i.e. 157 (defined via CHIMES_NETWORK_SIZE in cooling_struct.h). However, at run time the user can switch off individual metals from the network. In this case, the unused species will just be zeroed out at the end of the abundance array. Ideally, we would want to set the size of the array in the snapshot at run time. To do this, we would need to access the cooling_function_data structure within cooling_write_particles(), as this is where the actual network size is stored. However, when this routine is called in common_io.c, it is passed a NULL pointer for this argument, which means that we cannot access this structure.