Update to CHIMES parameters.
Some of the parameters in the CHIMES module are stored with a custom ChimesFloat type, which is set to double by default. When reading in the parameters from the parameter file in cooling_init_backend(), we previously added a check that ChimesFloat is indeed set to double, so that we could use the parser_get_param_double() routine, otherwise it would raise an error. I have now removed this check, and simply cast the parameters read in as doubles to ChimesFloat. This means that we can change the CHIMES module to single precision, simply be re-defining ChimesFloat (although in practice it is unlikely that we will ever use this option).