Added separate dust_ratio variable in CHIMES.
Previously, we just assumed that the dust_ratio, relative to the Milky Way dust abundance, just scales linearly with metallicity. However, in some applications we want to estimate the dust ratio more accurately, for example based on dust depletion models (as is done in the COLIBRE cooling tables). In this case, we need separate variables for the dust_ratio and the metallicity in CHIMES.