isotropic_photon_density read in from cross sections tables.
When UV_field_flag == 1, the isotropic_photon_density variable, which defines the normalisation of the radiation field, is now read in from the cross sections table, rather than having to specify it separately in the parameter file. We have also added an optional parameter, radiation_field_normalisation_factor, that re-scales the isotropic_photon_density from the table (default value: 1.0).