Implemented COLIBRE ISRF + shielding in CHIMES.
When UV_field_flag == 2, we set the radiation field as the sum of two spectra:

Firstly, the extra-galactic UVB. The user specifies the path to the UVB cross sections table, and so is free to choose the UVB (for consistency with the COLIBRE cooling tables, we should use Haardt & Madau 2012 here). However, the UVB is currently held fixed throughout the simulation. We will need to make this redshift-dependent before we start cosmological runs.

Secondly, the ISRF. The user specifies the path to the cross sections table to be used for the ISRF (for consistency with the COLIBRE cooling tables, we should use the Black 1987 ISRF here). The normalisation is then scaled based on the Jeans column density, in the same way as in the COLIBRE cooling tables. The cosmic ray ionisation rate and dust ratio are also scaled in this way.

When Shielding_flag == 2, we set the shielding length in CHIMES according to the shielding column density as used in the COLIBRE cooling tables.