Added rapid cooling switch in COLIBRE cooling.
We have added a new optional parameter, rapid_cooling_threshold (default value: -1), in the COLIBRE cooling mode that allows us to control how we implement the cooling, based on the ratio dt / t_cool.

If (dt / t_cool) > rapid_cooling_threshold, we are in the rapid cooling regime. Then, in cooling_cool_part() routine, we immediately update the particle's internal energy with the final internal energy at the end of the time-step, u_final, and we set du/dt = 0 so that the internal energy will not be drifted.

If (dt / t_cool) <= rapid_cooling_threshold, we are in the slow cooling regime. Then, in the cooling_cool_part() routine, we just set du/dt based on the overall change delta_u over the time-step dt. This is then used to drift the internal energy, as usual.

If rapid_cooling_threshold is negative, we always use the slow cooling regime method.

Note that we have also modified how the du/dt is set in the drift method (for the slow cooling regime), to make it more similar to the EAGLE cooling implementation. In particular, we no longer include the strict limits on du/dt within cooling_cool_part(), as the entropy floor and minimal internal energy are also imposed in the drift operations anyway.