Added calculate_colibre_N_ref() routine.
This routine calculates the column density, N_ref, that is used in COLIBRE to scale the ISRF, cosmic rays, dust depletion, and shielding length. We were previously calculating this within chimes_update_gas_vars(), but by moving it into a separate routine we can easily use this elsewhere as well, without needing to repeat the same code.