Implemented COLIBRE dust depletion factors into CHIMES.
In the COLIBRE cooling tables, the abundances of certain metals are reduced based on the dust depletion factors from Table 4 of Jenkins (2009). These dust depletion factors are then scaled with (Nref / NH0) ** 1.4 for Nref < NH0, constant otherwise.

I have added a new parameter in the CHIMES cooling module, colibre_metal_depletion. When this parameter is set to 1, we reduce the metal abundances due to dust depletion in the same way as in the COLIBRE cooling tables.