Implemented entropy floor in the CHIMES cooling module.
If a particle is within a certain factor of the EOS, controlled by the delta_logUEOS_apply_eqm parameter, then the chemistry and cooling are evolved in chemical equilibrium. This is achieved by looking up the equilibrium abundances from an HDF5 table, for a given temperature, density and metallicity, which are used to compute the cooling rate. This is then integrated over the time step using bisection integration scheme, similar to that used in the COLIBRE cooling module.