Added option to update ISRF with CHIMES rate equations.
When we use the COLIBRE ISRF (with UV_field_flag == 2), the strength of the radiation field is a function of density and temperature. This means that, as we integrate the chemistry and cooling, we need to update the radiation field within the rate equations.

We have added a new parameter, update_colibre_ISRF. If this is set to 1, then we re-compute the COLIBRE ISRF within the CHIMES rate equations. This is left as a parameter for now, so that we can compare with and without this option.