Updated the N_min parameter in the COLIBRE model used in CHIMES.
When we calculate the reference column density, N_ref, used in the COLIBRE model for the ISRF, cosmic rays, dust depletion and shielding, we previously used a value of N_min based on the max_shielding_length parameter. However, the user can set this parameter to be negative, to avoid imposing a maximum shielding length. In this case, N_min would be negative, but we then take log(N_min). I have therefore changed this so that N_min is hard-coded to a value of 3.08567758e15, which is the value used in Ploeckinger et al. (in prep.).

I have also added the const keyword for some of the variables used in calculating N_ref, where they have been hard-coded to the values in Ploeckinger et al. (in prep.).