Added option to update shielding length with CHIMES rate equations.
When we use the COLIBRE Shielding length (i.e. Shielding_flag == 2), the shielding length is a function of temperature and density. This means that, as we integrate the chemistry and cooling, we need to update the shielding length within the rate equations.

We have added a new parameter, update_colibre_shielding. If this is set to one then we re-compute the shielding length and resulting shielding column densities within the CHIMES rate equations.