Re-compute cosmic ray rate when update_colibre_ISRF == 1.
When we use the COLIBRE ISRF, the cosmic ray ionisation rate is also scaled with the reference column density, N_ref, and so is a function of temperature and density. If update_colibre_ISRF is set to 1, we now re-compute the cosmic ray rate within the CHIMES rate equations, in addition to the ISRF itself.