Implemented rapid cooling scheme in CHIMES.
This is the same scheme for updating a particle's internal energy from the cooling routines as was recently implemented in the COLIBRE cooling module.

There is a new parameter, rapid_cooling_threshold, in the CHIMES cooling mode that controls how the internal energy will be updated, based on the ratio dt / t_cool, i.e. the time-step over the cooling time.

If (dt / t_cool) >= rapid_cooling_threshold, we are in the rapid cooling regime. Then, in cooling_cool_part() routine, we immediately update the particle's internal energy with the final internal energy at the end of the time-step, u_final, and we set du/dt = 0 so that the internal energy will not be drifted.

If (dt / t_cool) < rapid_cooling_threshold, we are in the slow cooling regime. Then, in the cooling_cool_part() routine, we just set du/dt based on the overall change delta_u over the time-step dt. This is then used to drift the internal energy, as usual.

If rapid_cooling_threshold is negative, we always use the slow cooling regime method.