Update to the InitIonState parameter in CHIMES.
I have moved this parameter from ChimesGlobalVars to ChimesGasVars. Currently, this parameter is only used when setting the initial CHIMES abundances at the beginning of the simulation. However, when we use HII regions, we will also need to manually set the HII region particles to be fully ionised. We can use existing CHIMES routines for this, but by moving InitIonState to ChimesGasVars, it will be easier to specify different values for this parameter.