Update to HII regions in CHIMES.
For gas particles flagged as HII regions, after integrating the cooling we now manually set the particle to be ionised before calculating the final internal energy, u_final. Previously, we were doing this after calculating u_final, which meant that the latter was calculated using the wrong mean molecular weight.