Update to the entropy floor used with CHIMES.
If a particle reaches the entropy floor at the end of the time-step but was evolved with non-equilibrium chemistry (meaning that it started from a long way above the EOS), we now re-set its abundance array to be in chemical equilibrium on the floor. Previously, the abundance array would have been left with the non-eq abundances, and would not be updated again until the particle was next active.