Added parameters for hybrid cooling in CHIMES.
We have added a parameter, hybrid_cooling_mode, to control whether we use the hybrid cooling in CHIMES. If this mode is set to 0, we only use the CHIMES network for cooling. If it is set to 1, then we only use CHIMES for Hydrogen and Helium. Metal cooling will then be read in from the COLIBRE cooling tables within the CHIMES cooling routines (although this has not yet been implemented).

The COLIBRE cooling tables are now read in when hybrid_cooling_mode == 1. Also, in this case the various Include<Element> parameters for CHIMES are hard-coded to zero, and are not read in from the parameter file.

We have also fixed some minor bugs in how the InitIonState parameter is set when we set the initial abundance array at the start of the simulation. This parameter was not being correctly passed to the ChimesGasVars structure.