Update to hybrid cooling in CHIMES.
To switch on the Hybrid cooling in CHIMES, we now need to use the --with-cooling=CHIMES-HYBRID configure option. This will then set the hybrid_cooling_mode parameter in the ChimesGlobalVars structure, rather than reading this in from the parameter file.

By keeping this parameter in ChimesGlobalVars, we can in most cases determine whether we are using hybrid cooling from the parameter, without needing to use #ifdef's. However, when we set the size of the abundance array in xpart, we need to know this at compile time. If we instead wait until run time to determine the size of the abundance array, based solely on parameters read in from the parameter file, then we would need to make this a dynamically allocated array in xpart. However, this makes it complicated to move the particle data around, for example if we move a particle between MPI tasks.

By having a configure flag to determine when we use hybrid cooling, we can then define the size of the abundance array at compile time to be 10 if using hybrid cooling, and 157 otherwise. This is more efficient than always setting the size to be 157 and only using the first 10 elements of the array, as we were doing previously.