Added parameters for the COLIBRE ISRF/shielding models in CHIMES.
The N_ref column density in the COLIBRE model, which is used to scale the ISRF, shielding, depletion etc., has a high-temperature cut off that is controlled by the log_T_min and log_T_max parameters. These are now read in from the parameter file, so that we can explore how varying this cut-off affects the galaxy evolution. Note that this is for testing purposes only. Once we have decided on a model, these should be hard-coded to match what is used in the COLIBRE cooling tables.