Added cooling_convert_quantities() routine.
This routine is analogous to the hydro_convert_quantities() routine, and is run once at the end of the engine_init_particle() routine at the start of the calculation, but after the particle densities have been computed. This can be used to compute quantities for the cooling module for each particle at the beginning of the calculation that require the density.

Currently, this function does not do anything. It has been added as an empty "dummy routine" for all cooling modules. In the future, it will be used in the CHIMES cooling module for computing the initial equilibrium ion and molecule molecules at the start of the calculation.