Fixed bug in computing initial equilibrium abundances in CHIMES.
When the init_abundance_mode parameter is set to 2, we compute the initial equilibrium ion and molecule abundances for each gas particle at the beginning of the simulation. However, the way that this was being done previously meant that this was done before the particle densities were computed, which would lead to undefined behaviour.

I have now moved the calculation of the initial equilibrium abundances to inside the cooling_convert_quantities() routine. This is called at the end of the engine_init_particle() routine, but after the densities have been computed.