Added parameter to scale COLIBRE ISRF.
The ISRF, and also the cosmic ray ionisation rate, used in COLIBRE are reduced by a factor of ten in the fiducial COLIBRE cooling tables, compared to the Milky Way values. I have now made this a free parameter in CHIMES, controlled by the radiation_field_normalisation_factor parameter. Note that this only controls the ISRF and cosmic ray rate used by the CHIMES chemistry solver. If you are using the CHIMES-HYBRID cooling mode, it is up to the user to select the appropriate COLIBRE cooling table for the metal cooling and heating rates.