Fixed bug in the redshift-dependent UVB in CHIMES.
The cooling_init_backend() routine did not set the redshift in ChimesGlobalVars before calling init_chimes(). This meant that, when init_chimes() first sets up the redshift-dependent UVB, it is undefined as to what redshift it will set it to. If it defaults to redshift zero, this will cause problems later on when the redshift is set properly, as the UVB routines assume that the redshift is always decreasing. They cannot cope with an increase in the redshift.

The cooling_init_backend() routine does not have access to the initial redshift of the simulation, as it is not passed the cosmo structure. I have therefore set it up to initialise ChimesGlobalVars.redshift = 1000.0. This is much greater than the highest redshift bin used in the data tables, which ensures that, once the actual redshift is set later on in cooling_update(), it will not try to move the UVB to a higher redshift bin.