Updated the init_abundance_mode option in CHIMES.
Previously, setting this parameter to zero was supposed to read in the initial CHIMES abundances array from the ICs. However, this option was never actually implemented properly, and it is not practical to do this. I have therefore removed this option, and also added a new option so that the initial equilibrium abundances can be read in from a pre-computed equilibrium abundance table, rather than having to compute the initial equilibrium state for each particle.

I have also consolidated the code for setting the initial abundance array into the cooling_convert_quantities() routine. Previously, this was split between three different routines, depending on which mode was being used, but this made the code difficult to read.