Removed update_colibre_... options in CHIMES.
These options allowed us to update the ISRF, shielding column density and depletion factors used in the COLIBRE model within the CHIMES rate equations, as these quantities are temperature-dependent. However, this slowed down the chemistry solver by a factor ~2, as it required a lot more computations in every sub-cycle step. We found that this had no impact on the evolution or properties of the simulated galaxy, as these quantities still get updated every hydro time-step anyway. Also, it is arguably unphysical to update them within the rate equations in this way, as it is unrealistic to expect these quantities to evolve over the cooling time-scale. We therefore should not use these options.