Removed the colibre_log_T_min/max parameters in CHIMES.
These parameters control the high-temperature cut-off in the reference column density, N_ref, used to scale the ISRF, shielding column and dust depletion in the COLIBRE model. However, these should not be left as free parameters, as they need to be consistent with those used in the corresponding Cloudy cooling tables. We have therefore hard-coded these parameters to 3.0 and 5.0, respectively.