Fixed bug in CHIMES-Hybrid metal cooling rates.
When interpolating the metal cooling rates from the COLIBRE cooling tables, we were calling the interpolation4d_plus_summation() using colibre_cooling_N_internalenergy rather than colibre_cooling_N_temperature for the length of the axis of the table in the temperature direction. These differ by a factor of 2, so it would be accessing the wrong part of the table, and it could even access memory addresses outside the table altogether.