Fixed bug in setting initial CHIMES abundances.
When we compute the initial equilibrium CHIMES abundances in the cooling_convert_quantities() routine, we calculate the mean molecular weight from the chimes_abundances[] array in xp. However, at this point we had not copied over the initial guess for the abundance array to xp, as it was only set within the ChimesGasVars structure. This meant that the mean molecular weight, and hence the temperature, became inf.

We have now fixed this by immediately copying over the abundance array from ChimesGasVars to xp when we first set it within cooling_convert_quantities().