Fixed typos in colibre_cooling_rate_temperature()
The colibre_cooling_rate_temperature() routine interpolates the Tcooling, Theating and Telectron_fraction tables. However, in the calls to interpolation4d_plus_summation(), it was using colibre_cooling_N_internalenergy for the length of the table along the temperature axis, rather than colibre_cooling_N_temperature. As these differ by just over a factor of 2, this means that it would be accessing the wrong parts of the table here, and it could access memory addresses outside the table altogether. This has now been fixed.