Fixed bug in electron fractions for Hybrid cooling.
In the CHIMES-HYBRID cooling mode, the metal cooling rates from the COLIBRE tables are scaled according to the electron fraction. However, there was a mistake in how we were interpolating the primordial and metal electron fractions from the tables. We were sending the interpolation routine the abundance_ratios array for the weights, whereas we should be using a weight of unity for these two channels. This has now been fixed.

We also calculate the total metal electron fraction from the actual element abundances, to be used in the numerator of the electron_fraction_ratio.