Fixed bug in the abundance ratios used in the HYBRID cooling.
In the CHIMES-HYBRID cooling mode, the metal cooling and heating rates from the COLIBRE cooling tables are scaled according to the individual element abundances. However, if a particle has a non-zero metallicity that is lower than the lowest metallicity in the cooling tables, it uses the primordial bin, for which the log of the individual metal abundances are -31.93. This leads to the metal cooling rates being boosted by more than 20 orders of magnitude.

We have therefore changed the abundance_ratio_to_solar() routine such that, if a particle is using the primordial bin in the tables, then the ratio_solar[] array is set to zero for all metals. This ensures that the metal cooling and heating channels are zeroed out when using the primordial bins.