Update to how metallicity is set in CHIMES.
Previously, the metallicity that was passed to the CHIMES module was calculated as 1 - (XH + XHe). However, this can be affected by round off errors at very low metallicities. We have therefore changed this such that the metallicity is now calculated by summing over the individual metal element mass fractions. This ensures consistency with how the metallicity is calculated in the colibre tables routines that are used for the CHIMES-HYBRID cooling mode.