Rearranged how the abundance_ratio for hybrid cooling is passed to CHIMES.
We have added another hybrid_data struct within the ChimesGasVars structure, which contains the abundance_ratio array. This is defined within Swift and then passed to CHIMES as a void pointer. This allows us to define all of the data needed for the Swift-specific hybrid cooling function from Swift itself, without needing to add extra code within CHIMES.

We also added functions to allocate and free memory for this hybrid_data struct, which are called every time we allocate/free memory for the ChimesGasVars struct. These are again defined within Swift and passed to CHIMES as function pointers.