Added cooling_set_subgrid_properties() routine to CHIMES.
This routine sets the subgrid temperature and density in xp->tracers_data towards the end of cooling_cool_part(). If the particle is within dlogT_EOS of the EOS, we then re-compute the CHIMES abundance array to be in equilibrium using the subgrid density and temperature.

We have also added a new parameter, use_colibre_subgrid_EOS, to determine whether to use the subgrid density and temperature from the COLIBRE cooling tables. If this parameter is set to zero, the subgrid density and temperature are just set to the particle density and temperature.