Modified how the CHIMES abundances are computed on the subgrid EOS.
We have moved the calculation of the equilibrium CHIMES abundances for particles below logT_EOS_max into the cooling_set_subgrid_properties() routine. This will now set the subgrid density and subgrid temperature, and then set the equilibrium CHIMES abundances if required. This will make it easier for us to re-compute the equilibrium abundances when writing a snapshot.