Changed subgrid temperature and density of HII regions.
If a particle is flagged as an HII region but is on the EOS (i.e. below log10_T_EOS_max), we now limit the subgrid temperature to be no less than HIIregion_temp. If the subgrid temperature is limited in this way, the subgrid density is then set to be in pressure equilibrium at a temperature of HIIregion_temp.