Change to the subgrid hydrogen fractions for HII regions.
If a particle is flagged as an HII region, its HI, HII and H2 fractions are now always set according to the HIIregion_fion parameter. Previously, this was only done if the particle was above the EOS; for those HII regions below the EOS, the hydrogen fractions were being set at the thermal equilibrium temperature (at constant pressure).