Changes to the subgrid properties of HII regions in CHIMES.
In the CHIMES cooling module, if a particle is flagged as an HII region and lies below the EOS, its subgrid temperature is now limited to be no less than HIIregion_temp. The density is then in pressure equilibrium with this temperature.

Also, if an HII region particle lies above the EOS, its subgrid temperature is now set to the particle temperature, rather than HIIregion_temp.