Fixed the UVB around reionisation in CHIMES.
When interpolating the redshift-dependent UVB in CHIMES, if the two redshift bins that have been loaded bracket the reionisation redshift, we no longer interpolate the photoionisation cross sections and equilibrium abundances between the two bins. We instead just take one bin, either the lower or higher redshift, depending on whether the current redshift is after or before reionisation.

This fix avoids the spurious photoheating when using equilibrium cooling in CHIMES that we would otherwise get if we do not have a redshift bin exactly at the redshift of reionisation.