Added option for partially ionised HII regions in CHIMES.
The ionisation fraction of particles flagged as HII regions is now set by the HIIregion_ionization_fraction parameter, analogously to the COLIBRE cooling module. This parameter sets the fraction of each element that is in the singly ionised state, with the remainder set to neutral.

This parameter replaces the HIIregion_ion_state parameter.