Changes to CHIMES-HYBRID cooling.
There is no longer a separate --with-cooling configure option for CHIMES-HYBRID cooling. Instead, we just use the standard --with-cooling=CHIMES option, and then we also need to set the size of the CHIMES network in the configure script, as follows:

--with-chimes-network-size=10  --> CHIMES hybrid cooling (H+He only in non-equilibrium).
--with-chimes-network-size=157 --> Full CHIMES network (all elements in non-equilibrium).

We then need to specify whether to use hybrid cooling (i.e. to read in the metal cooling rates from the COLIBRE cooling tables) via the use_hybrid_cooling parameter.