Fixed subgrid properties of HII regions in CHIMES.
When we set the subgrid properties of gas particles in CHIMES, we forgot to check whether the particle is flagged as an HII region when setting the abundance array. This meant that, if it was below the EOS, its abundance array would be re-set to chemical equilibrium. We have now fixed this so that it first checks for HII regions, for which it sets the abundance array according to the HIIregion_fion parameter. If it is not an HII region, it will then check whether it is on/below the EOS.