Fixed subgrid properties on the EOS in CHIMES.
When setting the subgrid properties of EOS particles in CHIMES, we were using the maximum temperature above the EOS to determine whether a particle was considered as on the EOS. However, the cooling routine imposes the EOS as a floor on the internal energy rather than the temperature. This can lead to inconsistencies due to differences in the mean molecular weight. We have therefore changed the subgrid properties routines so that they check against the internal energy of the EOS, rather than the temperature, to ensure that this is all treated self consistently with other parts of the cooling routines.