Fixed errors from previous code formatting.
The format.sh re-arranged the order in which some of the header files were included in the CHIMES files. This led to compile errors, as some functions and data structures were no longer declared in the correct order. We have now re-ordered how we include the various header files in the CHIMES module so that it will compile correctly even after code formatting.