Fixed compile error in the CHIMES module.
When reading in the CHIMES cross sections tables in the CHIMES initialisation routines, we copy the path string for the filename into a string buffer. However, this buffer could potentially be too small for the string that we are passing to it. This can cause the compiler to exit with an error message.

We have therefore increased the size of this buffer, which ensures that the string will never exceed the buffer size.