Re-organised the CHIMES header files.
We have re-organised the CHIMES header files such that chimes_vars.h is only included within the CHIMES part of the code, and is not visible throughout the rest of the SWIFT code. Note that chimes_proto.h is still visible outside CHIMES and the cooling directory, because it contains structures and functions that are needed by cooling_struct.h and cooling.h. However, this at least means that the variables such as PI and the physical constants that are defined in chimes_vars.h are no longer visible outside CHIMES.