Fixed the explicit solution method in CHIMES.
When we take the explicit solution for the chemistry and cooling integration in CHIMES, we were not imposing the constraint equations on the final abundance array. This led to some abundances becoming slightly negative. We have now added a call to check_constraint_equations() after taking the explicit solution, which ensures that abundances can never be negative.