Fixed bug in how CHIMES treats particles on the entropy floor.
When a particle is on (or just above) the entropy floor, we set the abundances to chemical equilibrium. However, when the chimes_update_gas_vars() routine checked whether the particle is on the entropy floor, it was comparing its internal energy in cgs units to the entropy floor's internal energy in code units. This meant that particles were never switched over to using equilibrium cooling, and so they would always be evolved in non-equilibrium and then placed back on to the entropy floor at the end of the time-step. This has now been fixed.