Updates to the CHIMES documentation.
We have updated the CHIMES RTD documentation describing how to configure SWIFT with the CHIMES module, to reflect recent changes to how we define the size of the CHIMES network through the configure script.

We have also updated the error message that is given if the network size given in the configure script does not match that derived from the CHIMES parameters in the parameter file.