Fix to hard-coded solar metallicity in CHIMES module.
The colibre_metal_cooling_rate_temperature() had hard-coded the solar metallicity to be 0.0129. We don't have access to the cooling_function structure here, which stores Zsol, but I now pass Zsol in via the hybrid_data structure. We set hybrid_data->Zsol = cooling->Zsol when we first initialise the cooling backend. This means that we don't have to pass the entire cooling_function structure in to colibre_metal_cooling_rate_temperature().