Added dummy cooling_update_feedback_particle() routine.
We have added a routine, cooling_update_feedback_particle(), that is called whenever a gas particle's internal energy is updated due to SNII or SNIa feedback in the COLIBRE feedback routines. It can be used to update the cooling properties of that particle, stored in xpj. For now, this is just dummy routine, i.e. it doesn't do anything. This routine has been added to the cooling.h header files in the CHIMES, COLIBRE and EAGLE cooling modules. If any other cooling modules are to be used with the COLIBRE feedback routines, this dummy routine will need to be added to them as well.