Set FB-heated particles to chemical eqm before writin snapshot.
In the cooling_set_subgrid_properties() routine, if a particle is not on or near the EOS we now check the heated_by_FB flag. If the particle has been heated by feedback, we set it to chemical equilibrium. This ensures that if a particle has been heated by feedback but has not yet gone through the cooling routines when we write out a snapshot, its abundance array will still be set to equilibrium.