Re-arranged CHIMES I/O.
When writing out the particle data in cooling_write_particles() with the CHIMES cooling module, we have re-arranged the output fields so that the ChimesAbundances field is written out first. This means that the cooling_set_subgrid_properties() routine will be called before writing out the Temperatures and SubgridTemperatures, which ensures that these temperatures are calculated using a mean molecular weight that is consistent with the CHIMES abundances that are written out to the snapshot.