Fixed bug in configure.ac when using CHIMES.
In configure.ac, we were defining HAVECHIMESCOOLING if $with_cooling = CHIMES. This works if we are using the default network, with only H + He. However, if we define a different network size using, e.g., --with-cooling=CHIMES_157, this would not be picked up in this AM_CONDITIONAL statement. This meant that HAVECHIMESCOOLING would not be defined and so it could not compile.

We have now fixed this so that it will define HAVECHIMESCOOLING even when we define a different network size.