Added option to build CHIMES with GNU extensions.
Added a configure option, --enable-chimes-gnu-source, that will enable CHIMES to use the exp10() function from the GNU extensions. If you try to use this option but exp10() is not available in the math library, it will exit with an error message. If you do not use this option, CHIMES will use its own definition for the chimes_exp10() function. We leave this as an option for the User because on some machines our own definition can be faster than the exp10() function provided from the GNU extensions.