Minor updates to the CHIMES module.
When reading in the equilibrium abundances from the pre-computed tables, we check that the tabulated log of the abundances are not less than log10(CHIMES_FLT_MIN).

We have also added print statements at the beginning of init_chimes() to print out the float precision of CHIMES and the Sundials library. This makes it easier to check that we are using the precision that we expect.