Removed user option to enable GNU sources in CHIMES by hand.
We now enable GNU extensions and use the native exp10() function in CHIMES whenever exp10() is automatically found to be present in the math library, otherwise we use our own exp10() definition. We no longer have a configuration option to allow the User to switch between the different exp10() definitions by hand.