Updated the CHIMES module to the latest version.
The latest version of CHIMES fixed some minor bugs. Most importantly, there was an error in how we were setting the molecular abundances to zero at high temperatures (above T_mol), as the corresponding elements were not being added on anywhere else. This meant that the constraint equations might not be correctly enforced before we start integrating the chemistry, which can lead to unphysical values in the abundance array.

In the latest version of CHIMES, we have fixed this such that, when we zero the molecular abundances, the constituent elements are added on in the neutral ionisation state such that the constraint equations are preserved.